+++ src/Pages/TodoPage/OtherTitle.jsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const OtherTitle = () => {
+  return <OtherTitleName>👀 다른 카뎃은 무엇을?</OtherTitleName>;
+};
+
+const OtherTitleName = styled.div`
+  font-size: 25px;
+  margin-bottom: 10px;
+`;
+
+export default OtherTitle;
