+++ src/Pages/TodoPage/TodoCadetList.jsx
@@ -1,54 +1,14 @@
+import WarningNoList from './WarningNoList';
+import CadetListCheckbox from './CadetListCheckbox';
+import CadetListItem from './CadetListItem';
+
 import styled from 'styled-components';
 
-import Checkbox from '@mui/material/Checkbox';
-
 const TodoCadetList = ({ item }) => {
-  const WarningSignNoList = () => {
-    return (
-      <WarningSignList>
-        <span style={{ color: 'gray' }}>등록된 할 일이 없습니다!</span>
-      </WarningSignList>
-    );
-  };
-
-  const CadetListCheckbox = ({ list }) => {
-    return (
-      <CadetListCheckboxes>
-        <Checkbox
-          checked={list.titleCheck}
-          size="small"
-          sx={{
-            '& .MuiSvgIcon-root': { fontSize: 15 },
-          }}
-          disabled
-        />
-      </CadetListCheckboxes>
-    );
-  };
-
-  const CadetListItem = ({ list }) => {
-    return (
-      <span>
-        {list.titleCheck === true ? (
-          <span
-            style={{
-              textDecorationLine: 'line-through',
-              color: 'gray',
-            }}
-          >
-            {list.title}
-          </span>
-        ) : (
-          <span>{list.title}</span>
-        )}
-      </span>
-    );
-  };
-
   return (
     <TodoCadetListBody>
       {item.todoDtoList.length === 0 ? (
-        <WarningSignNoList />
+        <WarningNoList />
       ) : (
         <TodoCadetListContainer>
           {item.todoDtoList.map((list, index) => (
@@ -62,6 +22,7 @@ const TodoCadetList = ({ item }) => {
     </TodoCadetListBody>
   );
 };
+
 const TodoCadetListBody = styled.div`
   overflow: auto;
   margin-top: 5%;
@@ -71,16 +32,6 @@ const TodoCadetListBody = styled.div`
 const TodoCadetListContainer = styled.div``;
 const TodoCadetListEntity = styled.div`
   display: table;
-`;
-const WarningSignList = styled.div`
-  text-align: center;
-  // vertical-align: middle;
-  margin-top: 80px;
-  margin-left: -5%;
-`;
-const CadetListCheckboxes = styled.span`
-  margin-left: -10px;
-  margin-bottom: -5px;
 `;
 
 export default TodoCadetList;
