+++ src/Pages/TodoPage/TodoCadetName.jsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+const TodoCadetName = ({ index, item }) => {
+  return <TodoCadetNames key={index}>{item.userName}</TodoCadetNames>;
+};
+
+const TodoCadetNames = styled.div`
+  margin-top: -5px;
+  font-size: 20px;
+`;
+
+export default TodoCadetName;
