+++ src/Pages/TodoPage/TodoInputForm.jsx
@@ -1,0 +1,51 @@
+import { format } from 'date-fns';
+
+import styled from 'styled-components';
+
+const TodoInputForm = ({ onSubmit, onChange, toDo, date }) => {
+  const today = new Date();
+
+  return (
+    <form onSubmit={onSubmit}>
+      <InputFormBody>
+        <InputFormInput
+          onChange={onChange}
+          value={toDo.content}
+          type="text"
+          placeholder="오늘 할 일을 입력하세요."
+        />
+        <InputFormButton
+          variant="contained"
+          onClick={onSubmit}
+          disabled={format(today, 'yyyy-MM-dd') !== format(date, 'yyyy-MM-dd')}
+        >
+          추가
+        </InputFormButton>
+      </InputFormBody>
+    </form>
+  );
+};
+
+const InputFormBody = styled.div`
+  width: 100%;
+`;
+const InputFormInput = styled.input`
+  border: 0px;
+  border-bottom: 3px solid #c0c0c0;
+  margin-top: 30px;
+  font-size: 17px;
+  height: 35px;
+  width: calc(100% - 80px);
+  text-align: center;
+`;
+const InputFormButton = styled.button`
+  border-radius: 5px;
+  margin-left: 15px;
+  width: 60px;
+  height: 40px;
+  font-size: 17px;
+  background-color: transparent;
+  cursor: pointer;
+`;
+
+export default TodoInputForm;
