+++ src/Pages/TodoPage/TodoOtherList.jsx
@@ -1,12 +1,9 @@
+import TodoCadetList from './TodoCadetList';
+import TodoCadetName from './TodoCadetName';
+
 import styled from 'styled-components';
 
-import TodoCadetList from './TodoCadetList';
-
 const TodoOtherList = ({ othersToDo }) => {
-  const TodoCadetName = ({ index, item }) => {
-    return <TodoCadetNames key={index}>{item.userName}</TodoCadetNames>;
-  };
-
   return (
     <TodoOtherListBody key="index">
       {othersToDo.map((item, index) => (
@@ -34,10 +31,6 @@ const TodoOtherListContainer = styled.div`
   width: 100%;
   box-sizing: border-box;
   height: 300px;
-`;
-const TodoCadetNames = styled.div`
-  margin-top: -5px;
-  font-size: 20px;
 `;
 
 export default TodoOtherList;
